requirements.txt
aws-cdk.core
aws-cdk.aws-dynamodb
aws-cdk.aws-lambda
aws-cdk.aws_events
aws-cdk.aws_events_targets
boto3